requirements.txt
asgiref==3.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
diff-match-patch==20241021
Django==5.2.9
django-ckeditor==6.7.3
django-debug-toolbar==6.1.0
django-import-export==4.3.14
django-js-asset==3.1.2
django-modeltranslation==0.19.19
django-unfold==0.74.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==12.0.0
PyJWT==2.10.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
soupsieve==2.8
sqlparse==0.5.4
tablib==3.9.0
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
